Add basic support for static ips
The connman configuration is modified as explained in the following blog
post: https://resin.io/blog/how-to-configure-a-static-ip-address-with-resin-io/

Support for this feature is said to be "basic" since it misses advanced
settings like configuring a static ip for ethernet but not for wifi, or
having different static ip settings for certain network configurations.
More advanced configuration support will come in the future, since
deciding on a data structure that supports such flexibility would
introduce breaking changes in various Resin.io components, and thus
needs discussion.

The way it works now, if the network is configured to `ethernet`, the
static ip is applied to the ethernet section, and if the network `wifi`,
we configure the static ip for *both wifi and ethernet* sections.